update: remove duplicate sign in endpoint
