+++ src/app/api/v1/staffs/[staffId]/route.ts
@@ -82,7 +82,7 @@ export async function DELETE(req: NextRequest, { params }: paramsType) {
                 message: "success",
                 data: null,
             },
-            { status: 200 },
+            { status: 204 },
         );
     });
